+++ routes/api/users.js
@@ -22,22 +22,13 @@ const {
 } = require("../../middlewares/validation");
 const { auth } = require("../../middlewares/auth");
 
-router.post("/signup", registrationValidator, ctrlWrapper(userRegistration));
+router.post("/registration", registrationValidator, ctrlWrapper(userRegistration));
 router.post("/login", loginValidator, ctrlWrapper(userLogin));
 router.get("/current", auth, ctrlWrapper(userGetCurrent));
 router.get("/logout", auth, ctrlWrapper(userLogOut));
 router.delete("/:userId", auth, ctrlWrapper(userDelete));
-router.patch(
-  "/avatars",
-  auth,
-  upload.single("avatar"),
-  ctrlWrapper(userAvatar)
-);
+router.patch("/avatars", auth, upload.single("avatar"), ctrlWrapper(userAvatar));
 router.get("/verify/:verificationToken", ctrlWrapper(userVerification));
-router.post(
-  "/verify",
-  verificationValidator,
-  ctrlWrapper(userSendSecondEmail)
-);
+router.post("/verify", verificationValidator, ctrlWrapper(userSendSecondEmail));
 
 module.exports = router;
