+++ services/users/addToken.js
@@ -1,0 +1,8 @@
+const { User } = require("../../models/user");
+
+const addToken = async (email, token) => {
+    const result = await User.findOneAndUpdate({ email }, { token });
+    return result;
+  };
+
+  module.exports = { addToken }
